implement role based auth

+++ pages/api/deleteHoliday.ts
@@ -3,7 +3,8 @@ import { getSession } from "next-auth/react";
 
 export default async function handler(req, res) {
     const session = await getSession({ req });
-    if (session) {
+    // @ts-ignore
+    if (session && session.user.roll === "admin") {
         const deleteHolidays = await prisma.holidays.delete({
             where: {
                 id: parseInt(req.query.id),
